rm auth event
